separate the arrangement and the work
