+++ posthog/session_recordings/snapshots/serializer.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+
+class SessionRecordingSnapshotsSourceSerializer(serializers.Serializer):
+    source = serializers.CharField()  # type: ignore
+    start_timestamp = serializers.DateTimeField(allow_null=True)
+    end_timestamp = serializers.DateTimeField(allow_null=True)
+    blob_key = serializers.CharField(allow_null=True)
+
+
+class SessionRecordingSnapshotsSerializer(serializers.Serializer):
+    sources = serializers.ListField(child=SessionRecordingSnapshotsSourceSerializer(), required=False)
+    snapshots = serializers.ListField(required=False)
